Prep emojivoto for move to buoyantio/emojivoto repo
