+++ emojivoto-emoji-svc/emoji/emoji.go
