+++ emojivoto-emoji-svc/emoji/emoji_codemap.go
