+++ emojivoto-voting-svc/Makefile
