+++ emojivoto-voting-svc/api/api_test.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"testing"
 
-	pb "github.com/runconduit/conduit-examples/emojivoto/emojivoto-voting-svc/gen/proto"
-	"github.com/runconduit/conduit-examples/emojivoto/emojivoto-voting-svc/voting"
+	pb "github.com/buoyantio/emojivoto/emojivoto-voting-svc/gen/proto"
+	"github.com/buoyantio/emojivoto/emojivoto-voting-svc/voting"
 )
 
 func TestVoteJoy(t *testing.T) {
