+++ emojivoto-voting-svc/voting/poll.go
