+++ emojivoto-voting-svc/voting/poll_test.go
