+++ emojivoto-web/web/favicon.ico
