+++ emojivoto-web/webapp/js/components/Leaderboard.jsx
