+++ emojivoto-web/webapp/js/components/Vote.jsx
