+++ emojivoto-web/webapp/js/index.js
