+++ emojivoto-web/webapp/webpack.config.js
