--- lifecycle/bin/deploy
@@ -1,36 +1,0 @@
-#!/bin/sh
-
-set -e
-
-NAMESPACES=${1:-1}
-
-if [ "$NAMESPACES" -gt 100 ]; then
-  echo "Don't deploy more than 100 namespaces"
-  exit 1
-fi
-
-echo "Deploying $NAMESPACES injected, TLS, and baseline namespaces..."
-
-for i in `seq 1 $NAMESPACES`;
-do
-  NS=lifecycle$i
-  NS_TLS=$NS-tls
-  NS_BASELINE=$NS-baseline
-
-  echo "\nDeploying $NS..."
-  kubectl create ns $NS
-  cat lifecycle.yml |
-    conduit inject --conduit-namespace conduit-lifecycle - |
-    kubectl -n $NS apply -f -
-
-  echo "\nDeploying $NS_TLS..."
-  kubectl create ns $NS_TLS
-  cat lifecycle.yml |
-    conduit inject --conduit-namespace conduit-lifecycle-tls --tls optional - |
-    kubectl -n $NS_TLS apply -f -
-
-  echo "\nDeploying $NS_BASELINE..."
-  kubectl create ns $NS_BASELINE
-  cat lifecycle.yml |
-    kubectl -n $NS_BASELINE apply -f -
-done
