--- lifecycle/bin/teardown
@@ -1,18 +1,0 @@
-#!/bin/sh
-
-set -e
-
-NAMESPACES=${1:-1}
-
-echo "Tearing down $NAMESPACES injected, TLS, and baseline namespaces...\n"
-
-for i in `seq 1 $NAMESPACES`;
-do
-  NS=lifecycle$i
-  NS_TLS=$NS-tls
-  NS_BASELINE=$NS-baseline
-
-  kubectl delete ns $NS || true
-  kubectl delete ns $NS_TLS || true
-  kubectl delete ns $NS_BASELINE || true
-done
