+++ proto/Emoji.proto
