+++ proto/Voting.proto
